requirements.txt
# local package
-e .

# external requirements
invoke
click
Sphinx
coverage
#awscli
flake8
python-dotenv>=0.5.1

# libraries
python-dotenv
numpy
pandas
matplotlib
seaborn
statsmodels
scikit-learn
mlxtend
